Move auth-related models to ssas package
